# goa/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Pet Store Service
    description: HTTP service for managing pets
    version: ""
host: 127.0.0.1:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /pet:
        put:
            tags:
                - pets
            summary: CreatePet pets
            operationId: pets#CreatePet
            parameters:
                - name: CreatePetRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/PetsCreatePetRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
                "404":
                    description: Not Found response.
                    schema:
                        type: integer
                        format: int64
                "500":
                    description: Internal Server Error response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
    /pet/{id}:
        get:
            tags:
                - pets
            summary: PetByID pets
            operationId: pets#PetByID
            parameters:
                - name: id
                  in: path
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/PetsPetByIDOKResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/PetsPetByIDNotFoundResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PetsPetByIDInternalServerErrorResponseBody'
            schemes:
                - http
    /pets:
        get:
            tags:
                - pets
            summary: AllPets pets
            operationId: pets#AllPets
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/PetsPetResponseCollection'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PetsPetResponseCollection'
            schemes:
                - http
definitions:
    PetResponse:
        title: 'Mediatype identifier: pet; view=default'
        type: object
        properties:
            category:
                type: object
                properties:
                    id:
                        type: integer
                        example: 6886818317986174846
                        format: int64
                    name:
                        type: string
                        example: Optio praesentium repellat eum.
                example:
                    id: 3595954199181889644
                    name: Consequatur ea occaecati et voluptatum ab.
            id:
                type: integer
                example: 2289749508295316230
                format: int64
            name:
                type: string
                example: Explicabo nobis et.
            status:
                type: string
                example: Perferendis voluptatibus deleniti labore.
        description: PetResponse result type (default view)
        example:
            category:
                id: 8076905414576469655
                name: Voluptatibus unde odit odit.
            id: 4617109012261045417
            name: Quia repellendus architecto voluptate consequatur modi.
            status: Mollitia id odio.
    PetsCreatePetRequestBody:
        title: PetsCreatePetRequestBody
        type: object
        properties:
            category:
                type: object
                properties:
                    id:
                        type: integer
                        example: 4089838196721596704
                        format: int64
                    name:
                        type: string
                        example: At similique voluptas molestiae nisi dolor.
                example:
                    id: 6786498736116813843
                    name: Autem cupiditate nemo aut sint rerum ut.
            id:
                type: integer
                example: 9136542968667326800
                format: int64
            name:
                type: string
                example: Ea qui dolores.
            status:
                type: string
                example: Voluptas aut ut et doloribus est.
        example:
            category:
                id: 6754220346615845036
                name: Quis odio.
            id: 3357934641986696376
            name: Consequuntur quia iusto.
            status: Voluptas quae ut itaque odio laboriosam.
    PetsPetByIDInternalServerErrorResponseBody:
        title: 'Mediatype identifier: pet; view=default'
        type: object
        properties:
            category:
                type: object
                properties:
                    id:
                        type: integer
                        example: 1381978044063031361
                        format: int64
                    name:
                        type: string
                        example: Ipsam deserunt unde quibusdam reiciendis excepturi perferendis.
                example:
                    id: 8349416187893281922
                    name: Aut quos sit placeat.
            id:
                type: integer
                example: 3484453479257812748
                format: int64
            name:
                type: string
                example: Non inventore nobis aut fugiat velit.
            status:
                type: string
                example: Eligendi pariatur aspernatur iusto ut voluptas.
        description: PetByIDInternal Server ErrorResponseBody result type (default view)
        example:
            category:
                id: 4627490013532822212
                name: Ipsum temporibus est.
            id: 7135657107764564605
            name: Quaerat placeat praesentium voluptatum et ducimus.
            status: Velit modi modi.
    PetsPetByIDNotFoundResponseBody:
        title: 'Mediatype identifier: pet; view=default'
        type: object
        properties:
            category:
                type: object
                properties:
                    id:
                        type: integer
                        example: 8668015670417675875
                        format: int64
                    name:
                        type: string
                        example: Odio quaerat.
                example:
                    id: 7971434267246351389
                    name: Ab temporibus.
            id:
                type: integer
                example: 7357240065557550871
                format: int64
            name:
                type: string
                example: Illo necessitatibus sunt deleniti sed.
            status:
                type: string
                example: Nobis possimus suscipit modi odit nemo.
        description: PetByIDNot FoundResponseBody result type (default view)
        example:
            category:
                id: 2549732068515663792
                name: Voluptatem cupiditate qui ut.
            id: 4614544705709847434
            name: Fuga eos dolores beatae vero.
            status: Reiciendis natus.
    PetsPetByIDOKResponseBody:
        title: 'Mediatype identifier: pet; view=default'
        type: object
        properties:
            category:
                type: object
                properties:
                    id:
                        type: integer
                        example: 3648635734487108463
                        format: int64
                    name:
                        type: string
                        example: Veniam recusandae.
                example:
                    id: 8582094454689333535
                    name: Cupiditate nihil placeat ipsa.
            id:
                type: integer
                example: 1110157663399815708
                format: int64
            name:
                type: string
                example: Eum aliquam.
            status:
                type: string
                example: Fuga autem voluptas quia aperiam nulla.
        description: PetByIDOKResponseBody result type (default view)
        example:
            category:
                id: 5958051071156324538
                name: Beatae voluptatem nihil rerum libero perferendis odio.
            id: 7357210589769127070
            name: Sed adipisci exercitationem.
            status: Dolorum molestiae aspernatur.
    PetsPetResponseCollection:
        title: 'Mediatype identifier: pet; type=collection; view=default'
        type: array
        items:
            $ref: '#/definitions/PetResponse'
        description: AllPetsInternal Server ErrorResponseBody is the result type for an array of PetResponse (default view)
        example:
            - category:
                id: 4681448508605465902
                name: Esse et ut voluptatum ducimus.
              id: 6388154805178543352
              name: Placeat natus quisquam.
              status: Libero assumenda.
            - category:
                id: 4681448508605465902
                name: Esse et ut voluptatum ducimus.
              id: 6388154805178543352
              name: Placeat natus quisquam.
              status: Libero assumenda.
            - category:
                id: 4681448508605465902
                name: Esse et ut voluptatum ducimus.
              id: 6388154805178543352
              name: Placeat natus quisquam.
              status: Libero assumenda.
            - category:
                id: 4681448508605465902
                name: Esse et ut voluptatum ducimus.
              id: 6388154805178543352
              name: Placeat natus quisquam.
              status: Libero assumenda.